Add Teacher, fix id generator

+++ src/com/mebr0/user/base/Employee.java
@@ -4,7 +4,7 @@ import com.mebr0.user.util.IdGenerator;
 
 public abstract class Employee extends User {
 
-    public static int count = 0;
+    public static long count = 0;
 
     public Employee(String firstName, String lastName) {
         super(firstName, lastName);
